+++ pytorch_lightning/tuner/batch_size_scaling.py
@@ -164,8 +164,8 @@ def _run_power_scaling(
 
         if changed:
             # Force the train dataloader to reset as the batch size has changed
-            trainer.reset_train_dataloader(model)
-            trainer.reset_val_dataloader(model)
+            trainer._data_connector._reset_train_dataloader(model)
+            trainer._data_connector._reset_val_dataloader(model)
         else:
             break
     return new_size
@@ -203,8 +203,8 @@ def _run_binsearch_scaling(
 
             if changed:
                 # Force the train dataloader to reset as the batch size has changed
-                trainer.reset_train_dataloader(model)
-                trainer.reset_val_dataloader(model)
+                trainer._data_connector._reset_train_dataloader(model)
+                trainer._data_connector._reset_val_dataloader(model)
             else:
                 break