+++ tests/deprecated_api/test_remove_1-8.py
@@ -19,11 +19,12 @@ import torch
 from torch import optim
 
 from pytorch_lightning import Callback, Trainer
+from pytorch_lightning.trainer.states import RunningStage
 from pytorch_lightning.utilities import rank_zero_warn
 from pytorch_lightning.utilities.apply_func import move_data_to_device
 from pytorch_lightning.utilities.enums import DeviceType, DistributedType
 from pytorch_lightning.utilities.imports import _TORCHTEXT_LEGACY
-from tests.helpers.boring_model import BoringModel
+from tests.helpers.boring_model import BoringDataModule, BoringModel
 from tests.helpers.torchtext_utils import get_dummy_torchtext_data_iterator
 
 
@@ -256,3 +257,41 @@ def test_v1_8_0_deprecated_training_type_plugin_property():
     trainer = Trainer()
     with pytest.deprecated_call(match="in v1.6 and will be removed in v1.8"):
         trainer.training_type_plugin
+
+
+def test_v1_8_0_deprecate_trainer_data_loading_mixin():
+    trainer = Trainer(max_epochs=1)
+    model = BoringModel()
+    trainer.fit(model, datamodule=BoringDataModule())
+    reset_fit_methods = [
+        "reset_train_dataloader",
+        "reset_val_dataloader",
+        "reset_train_val_dataloaders",
+    ]
+    for method_name in reset_fit_methods:
+        with pytest.deprecated_call(
+            match=r"`TrainerDataLoadingMixin.%s` was deprecated in v1.6 and will be removed in v1.8." % method_name,
+        ):
+            fn = getattr(trainer, method_name, None)
+            fn()
+
+    trainer.test(model, datamodule=BoringDataModule())
+    with pytest.deprecated_call(
+        match=r"`TrainerDataLoadingMixin.reset_test_dataloader` was deprecated in v1.6 and will be removed in v1.8.",
+    ):
+        trainer.reset_test_dataloader()
+
+    trainer.predict(model, datamodule=BoringDataModule())
+    with pytest.deprecated_call(
+        match=r"`TrainerDataLoadingMixin.reset_predict_dataloader` was deprecated in v1.6 and will be removed in v1.8.",
+    ):
+        trainer.reset_predict_dataloader()
+
+    with pytest.deprecated_call(
+        match=r"`TrainerDataLoadingMixin.prepare_dataloader` was deprecated in v1.6 and will be removed in v1.8.",
+    ):
+        trainer.prepare_dataloader(dataloader=model.train_dataloader, shuffle=False)
+    with pytest.deprecated_call(
+        match=r"`TrainerDataLoadingMixin.request_dataloader` was deprecated in v1.6 and will be removed in v1.8.",
+    ):
+        trainer.request_dataloader(stage=RunningStage.PREDICTING)